+++ ja2lib/Tactical/TacticalSave.h
@@ -40,16 +40,16 @@ BOOLEAN SaveCurrentSectorsInformationToTempItemFile();
 BOOLEAN LoadCurrentSectorsInformationFromTempItemsFile();
 
 // Loads a World Item array from that sectors temp item file
-BOOLEAN LoadWorldItemsFromTempItemFile(INT16 sMapX, INT16 sMapY, INT8 bMapZ, WORLDITEM *pData);
+BOOLEAN LoadWorldItemsFromTempItemFile(u8 sMapX, u8 sMapY, INT8 bMapZ, WORLDITEM *pData);
 
 //  Adds an array of Item Objects to the specified location on a unloaded map.
 //  If you want to overwrite all the items in the array set fReplaceEntireFile to TRUE.
-BOOLEAN AddItemsToUnLoadedSector(INT16 sMapX, INT16 sMapY, INT8 bMapZ, INT16 sGridNo,
+BOOLEAN AddItemsToUnLoadedSector(u8 sMapX, u8 sMapY, INT8 bMapZ, INT16 sGridNo,
                                  UINT32 uiNumberOfItems, struct OBJECTTYPE *pObject, UINT8 ubLevel,
                                  UINT16 usFlags, INT8 bRenderZHeightAboveLevel, INT8 bVisible,
                                  BOOLEAN fReplaceEntireFile);
 
-BOOLEAN AddWorldItemsToUnLoadedSector(INT16 sMapX, INT16 sMapY, INT8 bMapZ, INT16 sGridNo,
+BOOLEAN AddWorldItemsToUnLoadedSector(u8 sMapX, u8 sMapY, INT8 bMapZ, INT16 sGridNo,
                                       UINT32 uiNumberOfItems, WORLDITEM *pWorldItem,
                                       BOOLEAN fOverWrite);
